+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Create Backup now also warns and requires typing `commit large files` to proceed if any changed/untracked file is 5 MB or larger (catches accidental `klippy.log`/firmware `.bin` commits that bloat repo size permanently).
+
 ## [v0.2.0] - 2026-07-14
 
 ### Added